+++ src/main/groovy/uk/q3c/krail/testutil/option/OptionDaoTestBase.groovy
@@ -24,11 +24,11 @@ import uk.q3c.krail.core.i18n.LabelKey
 import uk.q3c.krail.core.navigate.sitemap.comparator.DefaultUserSitemapSorters
 import uk.q3c.krail.core.option.*
 import uk.q3c.krail.core.persist.cache.option.OptionCacheKey
-import uk.q3c.krail.core.persist.clazz.i18n.ClassPatternSource
 import uk.q3c.krail.core.persist.common.option.OptionDao
 import uk.q3c.krail.core.persist.common.option.OptionSource
 import uk.q3c.krail.core.user.profile.UserHierarchy
 import uk.q3c.krail.core.view.component.LocaleContainer
+import uk.q3c.krail.i18n.api.clazz.ClassPatternSource
 
 import static uk.q3c.krail.core.user.profile.RankOption.*