+++ src/main/java/uk/q3c/krail/core/persist/common/common/KrailPersistenceUnitHelper.java
@@ -17,9 +17,9 @@ import com.google.inject.Binder;
 import com.google.inject.Provider;
 import com.google.inject.TypeLiteral;
 import com.google.inject.multibindings.MapBinder;
-import uk.q3c.krail.core.persist.common.i18n.PatternDaoProviders;
 import uk.q3c.krail.core.persist.common.option.OptionDaoDelegate;
 import uk.q3c.krail.i18n.api.PatternDao;
+import uk.q3c.krail.i18n.api.PatternDaoProviders;
 
 import java.lang.annotation.Annotation;