+++ src/main/java/uk/q3c/krail/core/persist/inmemory/common/InMemoryModule.java
@@ -20,7 +20,6 @@ import uk.q3c.krail.core.i18n.DescriptionKey;
 import uk.q3c.krail.core.i18n.LabelKey;
 import uk.q3c.krail.core.option.InMemory;
 import uk.q3c.krail.core.persist.common.common.*;
-import uk.q3c.krail.core.persist.common.i18n.PatternDaoProviders;
 import uk.q3c.krail.core.persist.common.option.OptionContainerProvider;
 import uk.q3c.krail.core.persist.common.option.OptionDaoDelegate;
 import uk.q3c.krail.core.persist.inmemory.i18n.DefaultInMemoryPatternStore;
@@ -29,6 +28,7 @@ import uk.q3c.krail.core.persist.inmemory.i18n.InMemoryPatternStore;
 import uk.q3c.krail.core.persist.inmemory.option.*;
 import uk.q3c.krail.i18n.api.I18NKey;
 import uk.q3c.krail.i18n.api.PatternDao;
+import uk.q3c.krail.i18n.api.PatternDaoProviders;
 
 import java.lang.annotation.Annotation;