+++ src/test/groovy/uk/q3c/krail/core/data/AnnotationOptionListConverterTest.groovy
@@ -17,8 +17,8 @@ import org.apache.commons.collections15.ListUtils
 import spock.lang.Specification
 import uk.q3c.krail.core.eventbus.SessionBus
 import uk.q3c.krail.core.option.AnnotationOptionList
-import uk.q3c.krail.core.persist.clazz.i18n.ClassPatternSource
 import uk.q3c.krail.i18n.VaadinCurrentLocale
+import uk.q3c.krail.i18n.api.clazz.ClassPatternSource
 
 /**
  *
@@ -58,7 +58,7 @@ class AnnotationOptionListConverterTest extends Specification {
 
         then:
 
-        string.equals("uk.q3c.krail.core.persist.clazz.i18n.ClassPatternSource~~uk.q3c.krail.core.eventbus.SessionBus")
+        string.equals("uk.q3c.krail.i18n.api.clazz.ClassPatternSource~~uk.q3c.krail.core.eventbus.SessionBus")
         ListUtils.isEqualList(testList.getList(), returnedList.getList())
     }