+++ src/test/groovy/uk/q3c/krail/i18n/DefaultPatternUtilityTest.groovy
@@ -18,11 +18,11 @@ import spock.lang.Specification
 import uk.q3c.krail.core.i18n.DescriptionKey
 import uk.q3c.krail.core.i18n.LabelKey
 import uk.q3c.krail.core.i18n.MessageKey
-import uk.q3c.krail.core.persist.clazz.i18n.ClassPatternDao
-import uk.q3c.krail.core.persist.clazz.i18n.ClassPatternSource
 import uk.q3c.krail.core.validation.ValidationKey
 import uk.q3c.krail.i18n.api.I18NKey
 import uk.q3c.krail.i18n.api.PatternSource
+import uk.q3c.krail.i18n.api.clazz.ClassPatternDao
+import uk.q3c.krail.i18n.api.clazz.ClassPatternSource
 
 //((ValidationKey.getEnumConstants().length + LabelKey.getEnumConstants().length + DescriptionKey.getEnumConstants().length + MessageKey.getEnumConstants().length) * 2) * patternDao.write(_, 'anything')
 /**