+++ src/test/groovy/uk/q3c/krail/i18n/VaadinI18NModuleTest.groovy
@@ -24,12 +24,12 @@ import uk.q3c.krail.core.guice.vsscope.VaadinSessionScopeModule
 import uk.q3c.krail.core.i18n.DescriptionKey
 import uk.q3c.krail.core.i18n.LabelKey
 import uk.q3c.krail.core.option.InMemory
-import uk.q3c.krail.core.persist.clazz.i18n.ClassPatternDao
-import uk.q3c.krail.core.persist.clazz.i18n.ClassPatternSource
 import uk.q3c.krail.core.persist.inmemory.common.InMemoryModule
 import uk.q3c.krail.core.persist.inmemory.i18n.InMemoryPatternDao
 import uk.q3c.krail.core.shiro.DefaultShiroModule
 import uk.q3c.krail.i18n.api.*
+import uk.q3c.krail.i18n.api.clazz.ClassPatternDao
+import uk.q3c.krail.i18n.api.clazz.ClassPatternSource
 import uk.q3c.krail.i18n.i18nModule.TestPatternSource
 import uk.q3c.krail.testutil.option.TestOptionModule