+++ Release Notes.md
@@ -11,5 +11,7 @@ There will, regrettably be quite a few changes in Krail apps, but with one or tw
 -`uk.q3c.krail.i18n.api.EnumResourceBundle` moved to **krail-i18n**
 - `uk.q3c.util.MessageFormat` moved to `uk.q3c.krail.i18n.MessageFormat`
 - `I18NModule` split, new `I18NModule` in **krail-i18n**, `VaadinI18NModule` in core
+- contents of `uk.q3c.krail.core.persist.clazz.i18n` moved to `uk.q3c.krail.i18n.clazz` or `uk.q3c.krail.i18n.api.clazz`
+- contents of `uk.q3c.krail.core.persist.common.i18n` moved to `uk.q3c.krail.i18n` or `uk.q3c.krail.i18n.api`